Fixing more memory erros, refactor some code and norminette them
