+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/13 17:22:12 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/13 12:30:28 by axlee            ###   ########.fr       */
+/*   Updated: 2024/06/17 15:10:02 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
