+++ libft/libft/str/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/13 17:53:03 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/10 12:03:59 by axlee            ###   ########.fr       */
+/*   Updated: 2024/06/17 15:10:55 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
