+++ output.txt
@@ -1,1 +1,1 @@
-hello , axlee
+BYE BYE BYE NSYNC
