+++ src/free_shell.c
@@ -1,34 +1,48 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   builtin_exit.c                                     :+:      :+:    :+:   */
+/*   free_shell.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/04/12 11:06:17 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/17 15:37:30 by axlee            ###   ########.fr       */
+/*   Created: 2024/06/17 15:22:18 by axlee             #+#    #+#             */
+/*   Updated: 2024/06/17 15:30:21 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void	minishell_exit(t_shell *minishell)
+void	free_env(char **env)
 {
 	int	i;
 
 	i = 0;
-	while (i < minishell->process_count)
+	if (env)
 	{
-		kill(minishell->process_ids[i], SIGTERM);
-		waitpid(minishell->process_ids[i], NULL, WNOHANG);
-		i++;
+		while (env[i])
+		{
+			free(env[i]);
+			i++;
+		}
+		free(env);
 	}
-	if (minishell->cmd_list != NULL)
+}
+
+void	free_shell(t_shell *minishell)
+{
+	if (minishell == NULL)
+		return ;
+	free_env(minishell->env);
+	if (minishell->cmd_list)
 	{
 		free_tokenlst(minishell->cmd_list);
 		minishell->cmd_list = NULL;
 	}
-	free_shell(minishell);
-	ft_putstr_fd("exit\n", STDOUT_FILENO);
-	exit(0);
+	if (minishell->input_fd != -1)
+		close(minishell->input_fd);
+	if (minishell->output_fd != -1)
+		close(minishell->output_fd);
+	if (minishell->prompt && ft_strcmp(minishell->prompt, PROMPT) != 0)
+		free(minishell->prompt);
+	free(minishell);
 }
