+++ src/parser/parse_double_quotes_utils_1.c
@@ -6,7 +6,7 @@
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/12 13:16:55 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/16 22:11:23 by axlee            ###   ########.fr       */
+/*   Updated: 2024/06/17 14:29:41 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,7 +71,6 @@ void	handle_variable_value(char **result, t_shell *minishell,
 	if (var_value)
 	{
 		append_to_result(result, minishell, var_value);
-		// minishell->flag = 1;
 		free(var_value);
 	}
 }
