+++ src/parser/parser.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parser.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gyong-si <gyong-si@student.42.fr>          +#+  +:+       +#+        */
+/*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/17 13:35:24 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/15 15:32:52 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/06/17 17:28:17 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,54 +31,7 @@ void	set_token_pointers(t_token *tokens)
 		prev_token->next = NULL;
 }
 
-static void	parse_unclosed(t_token *token, t_shell *minishell)
-{
-	char	*input;
-	char	*command;
-	int		inside_quote;
-	int		len;
-	int		i;
-	char	*trimmed_command;
-	char 	*delimiter;
-
-	i = 0;
-	input = token->token;
-	len = ft_strlen(input);
-	command = ft_strdup("");
-	inside_quote = 0;
-	minishell->flag = 0;
-	while (i < len || inside_quote)
-	{
-		if (i < len)
-		{
-			if (input[i] == '"')
-			{
-				inside_quote = !inside_quote;
-				command = ft_realloc(command, ft_strlen(command) + 2);
-				ft_strncat(command, &input[i], 1);
-			}
-		}
-		if (inside_quote && i == len - 1)
-		{
-			delimiter = "\"";
-			if (here_doc(minishell, delimiter, 2) == -1)
-			{
-				fprintf(stderr, "Error in here_doc\n");
-				free(command);
-				return ;
-			}
-			inside_quote = 0;
-		}
-		i++;
-	}
-	trimmed_command = ft_strtrim(command, "\"");
-	free(command);
-	free(token->token);
-	token->token = ft_strdup(trimmed_command);
-	free(trimmed_command);
-}
-
-static int	check_unclosed(t_token *token)
+int	check_unclosed(t_token *token)
 {
 	char	*str;
 	int		len;
@@ -98,6 +51,23 @@ static int	check_unclosed(t_token *token)
 	return (inside_quote);
 }
 
+void	handle_token_parsing(t_token *token, t_shell *minishell)
+{
+	char	*str;
+	int		len;
+
+	str = token->token;
+	len = ft_strlen(str);
+	if (str[0] == '\'' && str[len - 1] == '\'')
+		parse_single_quotes(token);
+	else if (str[0] == '\"' && str[len - 1] == '\"')
+		parse_double_quotes(token, minishell);
+	else if (check_unclosed(token) == 1)
+		parse_unclosed(token, minishell);
+	else
+		remove_embedded_quotes(token);
+}
+
 void	parse_token(t_token *token, t_shell *minishell)
 {
 	char	*str;
@@ -111,16 +81,7 @@ void	parse_token(t_token *token, t_shell *minishell)
 		token->token = ft_strdup("");
 	}
 	if (len > 1)
-	{
-		if (str[0] == '\'' && str[len - 1] == '\'')
-			parse_single_quotes(token);
-		else if (str[0] == '\"' && str[len - 1] == '\"')
-			parse_double_quotes(token, minishell);
-		else if (check_unclosed(token) == 1)
-			parse_unclosed(token, minishell);
-		else
-			remove_embedded_quotes(token);
-	}
+		handle_token_parsing(token, minishell);
 	str = token->token;
 	len = ft_strlen(str);
 	if (minishell->flag)
