+++ src/pipex/execute_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   execute_utils.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gyong-si <gyong-si@student.42.fr>          +#+  +:+       +#+        */
+/*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/06 10:21:39 by gyong-si          #+#    #+#             */
-/*   Updated: 2024/06/15 13:07:17 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/06/17 15:41:55 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
