+++ src/pipex/pipex_utils_3.c
@@ -1,36 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pipex_utils.c                                      :+:      :+:    :+:   */
+/*   pipex_utils_3.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/04/12 10:23:30 by gyong-si          #+#    #+#             */
-/*   Updated: 2024/06/14 15:25:44 by axlee            ###   ########.fr       */
+/*   Created: 2024/06/17 17:02:34 by axlee             #+#    #+#             */
+/*   Updated: 2024/06/17 17:04:01 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-void	exit_handler(int exit_code)
-{
-	if (exit_code == 1)
-		ft_putstr_fd("infile cmd cmd outfile\n", 2);
-	exit(EXIT_FAILURE);
-}
-
-int	open_file(const char *file, int mode)
-{
-	int	return_fd;
-
-	if (mode == 0)
-		return_fd = open(file, O_RDONLY, 0777);
-	else
-		return_fd = open(file, O_WRONLY | O_CREAT | O_TRUNC, 0777);
-	if (return_fd == -1)
-		exit_handler(1);
-	return (return_fd);
-}
 
 void	ft_free_tab(char **tab)
 {
@@ -45,7 +25,7 @@ void	ft_free_tab(char **tab)
 	free(tab);
 }
 
-//OPTION_1
+// OPTION_1
 /*static char	*try_exec_path(char **all_path, char **s_cmd)
 {
 	int		i;
@@ -86,8 +66,13 @@ char	*get_path(char *cmd, t_shell *minishell)
 	ft_free_tab(s_cmd);
 	return (exec);
 }*/
+void	ft_free_two_tabs(char **tab1, char **tab2)
+{
+	ft_free_tab(tab1);
+	ft_free_tab(tab2);
+}
 
-//OPTION_2 (PREFFERED)
+// OPTION_2 (PREFFERED)
 char	*get_path(char *cmd, t_shell *minishell)
 {
 	int		i;
@@ -107,18 +92,16 @@ char	*get_path(char *cmd, t_shell *minishell)
 		free(path_part);
 		if (access(exec, F_OK | X_OK) == 0)
 		{
-			ft_free_tab(s_cmd);
-			ft_free_tab(all_path);
+			ft_free_two_tabs(s_cmd, all_path);
 			return (exec);
 		}
 		free(exec);
 	}
-	ft_free_tab(all_path);
-	ft_free_tab(s_cmd);
+	ft_free_two_tabs(all_path, s_cmd);
 	return (NULL);
 }
 
-//OPTION_3
+// OPTION_3
 /*char	*get_path(char *cmd, t_shell *minishell)
 {
 	int		i;
@@ -147,17 +130,3 @@ char	*get_path(char *cmd, t_shell *minishell)
 	ft_free_tab(s_cmd);
 	return (NULL);
 }*/
-
-void	restore_fds(int input_fd, int output_fd)
-{
-	if (output_fd != STDOUT_FILENO)
-	{
-		dup2(output_fd, STDOUT_FILENO);
-		close(output_fd);
-	}
-	if (input_fd != STDIN_FILENO)
-	{
-		dup2(input_fd, STDIN_FILENO);
-		close(input_fd);
-	}
-}
